Refactor SAMPLE_HEADERS
