+++ src/frontend/src/views/Data/headers.tsx
@@ -1,88 +1,3 @@
-export const SAMPLE_HEADERS: Header[] = [
-  {
-    key: "privateId",
-    sortKey: ["privateId"],
-    text: "Private ID",
-    tooltip: {
-      boldText: "Private ID",
-      regularText:
-        "User-provided private ID. Only users in your Group can see it.",
-    },
-  },
-  {
-    key: "publicId",
-    sortKey: ["publicId"],
-    text: "Public ID",
-    tooltip: {
-      boldText: "Public ID",
-      regularText:
-        "This is your GISAID ID or public ID generated by CZ Gen Epi.",
-    },
-  },
-  {
-    key: "collectionDate",
-    sortKey: ["collectionDate"],
-    text: "Collection Date",
-    tooltip: {
-      boldText: "Collection Date",
-      regularText:
-        "User-provided date on which the sample was collected from an individual or an environment.",
-    },
-  },
-  {
-    key: "lineage",
-    sortKey: ["lineage", "lineage"],
-    text: "Lineage",
-    tooltip: {
-      boldText: "Lineage",
-      link: {
-        href: "https://cov-lineages.org/pangolin.html",
-        linkText: "Learn more.",
-      },
-      regularText:
-        "A lineage is a named group of related sequences. A few lineages have been associated with changes in the epidemiological or biological characteristics of the virus. We continually update these lineages based on the evolving Pangolin designations. Lineages determined by Pangolin.",
-    },
-  },
-  {
-    key: "uploadDate",
-    sortKey: ["uploadDate"],
-    text: "Upload Date",
-    tooltip: {
-      boldText: "Upload Date",
-      regularText: "Date on which the sample was uploaded to CZ Gen Epi.",
-    },
-  },
-  {
-    key: "collectionLocation",
-    sortKey: ["collectionLocation"],
-    text: "Collection Location",
-    tooltip: {
-      boldText: "Collection Location",
-      regularText:
-        "User-provided geographic location where the sample was collected (at the county level or above).",
-    },
-  },
-  {
-    key: "sequencingDate",
-    sortKey: ["sequencingDate"],
-    text: "Sequencing Date",
-    tooltip: {
-      boldText: "Sequencing Date",
-      regularText: "User-provided date on which the sample was sequenced.",
-    },
-  },
-  {
-    key: "gisaid",
-    sortKey: ["gisaid", "status"],
-    text: "GISAID",
-    tooltip: {
-      boldText: "GISAID Status",
-      regularText:
-        "Whether your sample has been Not Yet Submitted, Submitted, Accepted (with GISAID accession), Rejected, or Not Eligible (marked private).",
-    },
-  },
-];
-
 export const SAMPLE_SUBHEADERS: Record<string, SubHeader[]> = {
   gisaid: [
     {
@@ -140,6 +55,7 @@ export const SAMPLE_SUBHEADERS: Record<string, SubHeader[]> = {
   ],
 };
 
+// TODO: (ehoops) - we should refactor this config to follow the same pattern as the samples table
 export const TREE_HEADERS: Header[] = [
   {
     key: "name",
